moved promote and norm traits to xutils.hpp
